Move `ResolutionError` to the right folder
